# documents/api-specs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Match Real API",
    "description": "Match Real 서비스의 API 문서",
    "version": "v1.0.0"
  },
  "servers": [
    { "url": "https://api.matchreal.shop", "description": "Production" }
  ],
  "paths": {
    "/api/v1/login/refresh": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "4xx or 5xx": {
            "description": "데이터베이스 관련 예외 발생",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "4xx": {
            "description": "네트워크 관련 예외 발생",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "5xx": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "required": ["code", "errors", "message", "path", "timestamp"],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "에러 발생 시간",
            "format": "date-time"
          },
          "code": { "type": "string", "description": "에러 코드" },
          "message": { "type": "string", "description": "에러 메시지" },
          "path": {
            "type": "string",
            "description": "에러가 발생한 요청 경로"
          },
          "errors": {
            "type": "array",
            "description": "필드 에러 목록",
            "items": { "$ref": "#/components/schemas/FieldError" }
          }
        },
        "description": "에러 응답 객체"
      },
      "FieldError": {
        "required": ["field", "reason"],
        "type": "object",
        "properties": {
          "field": { "type": "string", "description": "에러가 발생한 필드명" },
          "value": {
            "type": "object",
            "description": "에러가 발생한 필드의 값"
          },
          "reason": { "type": "string", "description": "에러 이유" }
        },
        "description": "필드 에러 정보"
      },
      "RefreshTokenRequest": {
        "required": ["refreshToken"],
        "type": "object",
        "properties": { "refreshToken": { "type": "string" } }
      },
      "TokenResponse": {
        "required": ["accessToken", "refreshToken"],
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "description": "JWT 토큰을 입력해주세요.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}